Match the chat path through the shared pathname matcher
isAiChatPath needs a pathname rather than a router Location, since it is
called outside React from window.location, and it had grown its own copy
of the matchPath call. Both it and isMatchingLocation now share one
matcher.
